+++ sentry/src/main/java/io/sentry/ISentryClient.java
@@ -1,5 +1,6 @@
 package io.sentry;
 
+import io.sentry.protocol.Feedback;
 import io.sentry.protocol.Message;
 import io.sentry.protocol.SentryId;
 import io.sentry.protocol.SentryTransaction;
@@ -76,6 +77,32 @@ public interface ISentryClient {
    */
   default @NotNull SentryId captureEvent(@NotNull SentryEvent event, @Nullable Hint hint) {
     return captureEvent(event, null, hint);
+  }
+
+  /**
+   * Captures the feedback.
+   *
+   * @param feedback The feedback to send.
+   * @param hint An optional hint to be applied to the event.
+   * @param scope An optional scope to be applied to the event.
+   * @return The Id (SentryId object) of the event
+   */
+  default @NotNull SentryId captureFeedback(
+      @NotNull Feedback feedback, @Nullable Hint hint, @NotNull IScope scope) {
+    SentryEvent event = new SentryEvent();
+    event.getContexts().setFeedback(feedback);
+
+    if (feedback.getReplayId() == null) {
+      final @NotNull SentryId replayId = scope.getReplayId();
+      if (!replayId.equals(SentryId.EMPTY_ID)) {
+        feedback.setReplayId(replayId);
+      }
+    }
+    if (feedback.getUrl() == null) {
+      feedback.setUrl(scope.getScreen());
+    }
+
+    return captureEvent(event, scope, hint);
   }
 
   /**