+++ sentry/src/test/java/io/sentry/JsonSerializerTest.kt
@@ -3,6 +3,7 @@ package io.sentry
 import io.sentry.profilemeasurements.ProfileMeasurement
 import io.sentry.profilemeasurements.ProfileMeasurementValue
 import io.sentry.protocol.Device
+import io.sentry.protocol.Feedback
 import io.sentry.protocol.ReplayRecordingSerializationTest
 import io.sentry.protocol.Request
 import io.sentry.protocol.SdkVersion
@@ -1480,6 +1481,42 @@ class JsonSerializerTest {
             ),
             deserializedCollection
         )
+    }
+
+    @Test
+    fun `serializes feedback`() {
+        val replayId = SentryId("00000000-0000-0000-0000-000000000001")
+        val eventId = SentryId("00000000-0000-0000-0000-000000000002")
+        val feedback = Feedback("message")
+        feedback.name = "name"
+        feedback.contactEmail = "email"
+        feedback.url = "url"
+        feedback.setReplayId(replayId)
+        feedback.setAssociatedEventId(eventId)
+        val actual = serializeToString(feedback)
+        val expected = "{\"message\":\"message\",\"contact_email\":\"email\",\"name\":\"name\",\"associated_event_id\":\"00000000000000000000000000000002\",\"replay_id\":\"00000000000000000000000000000001\",\"url\":\"url\"}"
+        assertEquals(expected, actual)
+    }
+
+    @Test
+    fun `deserializes feedback`() {
+        val json = """{
+            "message":"message",
+            "contact_email":"email",
+            "name":"name",
+            "associated_event_id":"00000000000000000000000000000002",
+            "replay_id":"00000000000000000000000000000001",
+            "url":"url"
+        }"""
+        val feedback = fixture.serializer.deserialize(StringReader(json), Feedback::class.java)
+        val expected = Feedback("message").apply {
+            name = "name"
+            contactEmail = "email"
+            url = "url"
+            setReplayId(SentryId("00000000-0000-0000-0000-000000000001"))
+            setAssociatedEventId(SentryId("00000000-0000-0000-0000-000000000002"))
+        }
+        assertEquals(expected, feedback)
     }
 
     @Test